Patch: Fixed Addresource param to take int productionrate instead of float

+++ Game/Scripts/ResourceManager.h
@@ -12,7 +12,7 @@ public:
 	ResourceManager();
 	~ResourceManager();
 
-	Resources AddResource(ResourceName resName, string itemName, string SKU, int itemPrice, int itemStock, string resourceIcon, int stackLimit, int demand, float productionRate);
+	Resources AddResource(ResourceName resName, string itemName, string SKU, int itemPrice, int itemStock, string resourceIcon, int stackLimit, int demand, int productionRate);
 
 	/*!!! Stop using: Depricated !!!
 	!!! Stop using: Depricated !!!
